+++ gegelatilib/src/log/Logger.cpp
@@ -1,0 +1,9 @@
+#include <iostream>
+
+#include "log/Logger.h"
+
+Log::Logger Log::Logger::operator<<(std::ostream &(*manip)(std::ostream &)) {
+    manip(*out);
+    return *this;
+}
+